+++ src/content/mod.rs
@@ -14,4 +14,5 @@ pub use operators::{Operator, TextElement};
 pub use parser::parse_and_execute_text_only;
 pub use parser::parse_content_stream;
 pub use parser::parse_content_stream_images_only;
+pub use parser::parse_content_stream_paths_only;
 pub use parser::parse_content_stream_text_only;